requirements.txt
Pillow
torch
transformers
gradio
